requirements.txt

flask
numpy
pandas
scikit-learn
tensorflow
yfinance
nltk
transformers
dwave-ocean-sdk
plotly
newsapi-python
tweepy